@@ -1,6 +1,6 @@
 [project]
 name = "spin"
-requires-python = ">=3.9"
+requires-python = ">=3.10"  # Oldest non-EOL Python here
 description = "Developer tool for scientific Python libraries"
 readme = "README.md"
 license = {file = "LICENSE"}
@@ -15,7 +15,6 @@ classifiers = [
   "License :: OSI Approved :: BSD License",
   "Operating System :: OS Independent",
   "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3.9",
   "Programming Language :: Python :: 3.10",
   "Programming Language :: Python :: 3.11",
   "Programming Language :: Python :: 3.12",
